server/requirements.txt
aioredis==1.2.0
asgiref==3.2.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.9.2
Automat==0.7.0
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.39.1
django-cors-headers==3.0.2
django-extensions==2.1.9
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.0.7
djangorestframework==3.9.4
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
msgpack==0.6.1
oauthlib==3.0.1
openapi-codec==1.3.2
Pillow==6.1.0
psycopg2==2.8.3
pycparser==2.19
PyHamcrest==1.9.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.3
zope.interface==4.6.0
